+++ src/components/spartan/SSelector/index.ts
@@ -1,3 +1,3 @@
 export { default as SSelector } from './SSelector.vue';
 
-export type { TSelectorProps, TSelectorEmits } from './types';
+export type { TSelectorProps, TSelectorEmits, TSelectorSlots } from './types';